Connected back-end and front-end input data

+++ back_end/demo/src/main/java/com/back/example/InputSheet/BasicData.java
@@ -2,7 +2,7 @@ package com.back.example.InputSheet;
 
 public class BasicData {
 
-    /*public BasicData(String industry_Global_A7, float revenues_A8, float operating_ncome_or_EBIT_A9, float interest_expense_A10, float book_value_equity_A11, float book_value_debt_A12, boolean do_you_have_RD_expenses_to_capitalize_A13, boolean do_you_have_operating_lease_commitment_A14, float cash_and_Marketable_Securities_A15, float cross_holdings_and_other_nonOperating_assets_A16, float minority_interests_A17, float number_of_shares_outstanding_A18, float current_stock_price_A19, float effective_tax_rate_A20, float marginal_tax_rate_A21) {
+    /*public BasicData(String industry_Global_A7, double revenues_A8, double operating_ncome_or_EBIT_A9, double interest_expense_A10, double book_value_equity_A11, double book_value_debt_A12, boolean do_you_have_RD_expenses_to_capitalize_A13, boolean do_you_have_operating_lease_commitment_A14, double cash_and_Marketable_Securities_A15, double cross_holdings_and_other_nonOperating_assets_A16, double minority_interests_A17, double number_of_shares_outstanding_A18, double current_stock_price_A19, double effective_tax_rate_A20, double marginal_tax_rate_A21) {
         Industry_Global_A7 = industry_Global_A7;
         Revenues_A8 = revenues_A8;
         Operating_ncome_or_EBIT_A9 = operating_ncome_or_EBIT_A9;
@@ -28,43 +28,43 @@ public class BasicData {
         Industry_Global_A7 = industry_Global_A7;
     }
 
-    public float getRevenues_A8() {
+    public double getRevenues_A8() {
         return Revenues_A8;
     }
 
-    public void setRevenues_A8(float revenues_A8) {
+    public void setRevenues_A8(double revenues_A8) {
         Revenues_A8 = revenues_A8;
     }
 
-    public float getOperating_ncome_or_EBIT_A9() {
+    public double getOperating_ncome_or_EBIT_A9() {
         return Operating_ncome_or_EBIT_A9;
     }
 
-    public void setOperating_ncome_or_EBIT_A9(float operating_ncome_or_EBIT_A9) {
+    public void setOperating_ncome_or_EBIT_A9(double operating_ncome_or_EBIT_A9) {
         Operating_ncome_or_EBIT_A9 = operating_ncome_or_EBIT_A9;
     }
 
-    public float getInterest_expense_A10() {
+    public double getInterest_expense_A10() {
         return Interest_expense_A10;
     }
 
-    public void setInterest_expense_A10(float interest_expense_A10) {
+    public void setInterest_expense_A10(double interest_expense_A10) {
         Interest_expense_A10 = interest_expense_A10;
     }
 
-    public float getBook_value_equity_A11() {
+    public double getBook_value_equity_A11() {
         return Book_value_equity_A11;
     }
 
-    public void setBook_value_equity_A11(float book_value_equity_A11) {
+    public void setBook_value_equity_A11(double book_value_equity_A11) {
         Book_value_equity_A11 = book_value_equity_A11;
     }
 
-    public float getBook_value_debt_A12() {
+    public double getBook_value_debt_A12() {
         return Book_value_debt_A12;
     }
 
-    public void setBook_value_debt_A12(float book_value_debt_A12) {
+    public void setBook_value_debt_A12(double book_value_debt_A12) {
         Book_value_debt_A12 = book_value_debt_A12;
     }
 
@@ -84,77 +84,77 @@ public class BasicData {
         Do_you_have_operating_lease_commitment_A14 = do_you_have_operating_lease_commitment_A14;
     }
 
-    public float getCash_and_Marketable_Securities_A15() {
+    public double getCash_and_Marketable_Securities_A15() {
         return Cash_and_Marketable_Securities_A15;
     }
 
-    public void setCash_and_Marketable_Securities_A15(float cash_and_Marketable_Securities_A15) {
+    public void setCash_and_Marketable_Securities_A15(double cash_and_Marketable_Securities_A15) {
         Cash_and_Marketable_Securities_A15 = cash_and_Marketable_Securities_A15;
     }
 
-    public float getCross_holdings_and_other_nonOperating_assets_A16() {
+    public double getCross_holdings_and_other_nonOperating_assets_A16() {
         return Cross_holdings_and_other_nonOperating_assets_A16;
     }
 
-    public void setCross_holdings_and_other_nonOperating_assets_A16(float cross_holdings_and_other_nonOperating_assets_A16) {
+    public void setCross_holdings_and_other_nonOperating_assets_A16(double cross_holdings_and_other_nonOperating_assets_A16) {
         Cross_holdings_and_other_nonOperating_assets_A16 = cross_holdings_and_other_nonOperating_assets_A16;
     }
 
-    public float getMinority_interests_A17() {
+    public double getMinority_interests_A17() {
         return Minority_interests_A17;
     }
 
-    public void setMinority_interests_A17(float minority_interests_A17) {
+    public void setMinority_interests_A17(double minority_interests_A17) {
         Minority_interests_A17 = minority_interests_A17;
     }
 
-    public float getNumber_of_shares_outstanding_A18() {
+    public double getNumber_of_shares_outstanding_A18() {
         return Number_of_shares_outstanding_A18;
     }
 
-    public void setNumber_of_shares_outstanding_A18(float number_of_shares_outstanding_A18) {
+    public void setNumber_of_shares_outstanding_A18(double number_of_shares_outstanding_A18) {
         Number_of_shares_outstanding_A18 = number_of_shares_outstanding_A18;
     }
 
-    public float getCurrent_stock_price_A19() {
+    public double getCurrent_stock_price_A19() {
         return Current_stock_price_A19;
     }
 
-    public void setCurrent_stock_price_A19(float current_stock_price_A19) {
+    public void setCurrent_stock_price_A19(double current_stock_price_A19) {
         Current_stock_price_A19 = current_stock_price_A19;
     }
 
-    public float getEffective_tax_rate_A20() {
+    public double getEffective_tax_rate_A20() {
         return Effective_tax_rate_A20;
     }
 
-    public void setEffective_tax_rate_A20(float effective_tax_rate_A20) {
+    public void setEffective_tax_rate_A20(double effective_tax_rate_A20) {
         Effective_tax_rate_A20 = effective_tax_rate_A20;
     }
 
-    public float getMarginal_tax_rate_A21() {
+    public double getMarginal_tax_rate_A21() {
         return Marginal_tax_rate_A21;
     }
 
-    public void setMarginal_tax_rate_A21(float marginal_tax_rate_A21) {
+    public void setMarginal_tax_rate_A21(double marginal_tax_rate_A21) {
         Marginal_tax_rate_A21 = marginal_tax_rate_A21;
     }
 
     private String Industry_Global_A7;
-    private float Revenues_A8;
-    private float Operating_ncome_or_EBIT_A9;
-    private float Interest_expense_A10;
-    private float Book_value_equity_A11;
-    private float Book_value_debt_A12;
+    private double Revenues_A8;
+    private double Operating_ncome_or_EBIT_A9;
+    private double Interest_expense_A10;
+    private double Book_value_equity_A11;
+    private double Book_value_debt_A12;
     private boolean Do_you_have_RD_expenses_to_capitalize_A13;
     private boolean Do_you_have_operating_lease_commitment_A14;
-    private float Cash_and_Marketable_Securities_A15;
-    private float Cross_holdings_and_other_nonOperating_assets_A16;
-    private float Minority_interests_A17;
-    private float Number_of_shares_outstanding_A18;
-    private float Current_stock_price_A19;
-    private float Effective_tax_rate_A20;
-    private float Marginal_tax_rate_A21;
+    private double Cash_and_Marketable_Securities_A15;
+    private double Cross_holdings_and_other_nonOperating_assets_A16;
+    private double Minority_interests_A17;
+    private double Number_of_shares_outstanding_A18;
+    private double Current_stock_price_A19;
+    private double Effective_tax_rate_A20;
+    private double Marginal_tax_rate_A21;
 
 
 }
