+++ back_end/demo/src/main/java/com/back/example/InputSheet/DefaultAssumptions.java
@@ -1,7 +1,7 @@
 package com.back.example.InputSheet;
 
 public class DefaultAssumptions {
-/*    public DefaultAssumptions(boolean override_stable_growth_A41, float if_yes_enter_the_cost_of_capital_after_year10_A42, boolean override_competitive_advantages_will_fade_A44, float if_yes_enter_capital_you_expect_after_year10_A45, boolean override_firm_has_no_chance_of_failure_A47, float if_yes_enter_the_probability_of_failure_A48, String what_do_you_want_to_tie_your_proceeds_in_failure_to_A49, float enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50, boolean override_long_term_tax_rate_assumption_A52, boolean overrude_loss_tax_shield_assumption_A54, float if_yes_enter_NOL_carrying_into_year1_A55, boolean override_long_term_risk_free_assumptions_A57, float if_yes_enter_riskfree_rate_after_year10_A58, boolean override_long_term_growth_rate_assumptions_A60, float if_yes_enter_the_growth_rate_in_perpetuity_A61, boolean override_cash_is_a_neutral_asset_A63, float if_yes_enter_trapped_cash_or_entire_balance_A64, float average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65) {
+/*    public DefaultAssumptions(boolean override_stable_growth_A41, double if_yes_enter_the_cost_of_capital_after_year10_A42, boolean override_competitive_advantages_will_fade_A44, double if_yes_enter_capital_you_expect_after_year10_A45, boolean override_firm_has_no_chance_of_failure_A47, double if_yes_enter_the_probability_of_failure_A48, String what_do_you_want_to_tie_your_proceeds_in_failure_to_A49, double enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50, boolean override_long_term_tax_rate_assumption_A52, boolean overrude_loss_tax_shield_assumption_A54, double if_yes_enter_NOL_carrying_into_year1_A55, boolean override_long_term_risk_free_assumptions_A57, double if_yes_enter_riskfree_rate_after_year10_A58, boolean override_long_term_growth_rate_assumptions_A60, double if_yes_enter_the_growth_rate_in_perpetuity_A61, boolean override_cash_is_a_neutral_asset_A63, double if_yes_enter_trapped_cash_or_entire_balance_A64, double average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65) {
         Override_stable_growth_A41 = override_stable_growth_A41;
         If_yes_enter_the_cost_of_capital_after_year10_A42 = if_yes_enter_the_cost_of_capital_after_year10_A42;
         Override_competitive_advantages_will_fade_A44 = override_competitive_advantages_will_fade_A44;
@@ -30,11 +30,11 @@ public class DefaultAssumptions {
         Override_stable_growth_A41 = override_stable_growth_A41;
     }
 
-    public float getIf_yes_enter_the_cost_of_capital_after_year10_A42() {
+    public double getIf_yes_enter_the_cost_of_capital_after_year10_A42() {
         return If_yes_enter_the_cost_of_capital_after_year10_A42;
     }
 
-    public void setIf_yes_enter_the_cost_of_capital_after_year10_A42(float if_yes_enter_the_cost_of_capital_after_year10_A42) {
+    public void setIf_yes_enter_the_cost_of_capital_after_year10_A42(double if_yes_enter_the_cost_of_capital_after_year10_A42) {
         If_yes_enter_the_cost_of_capital_after_year10_A42 = if_yes_enter_the_cost_of_capital_after_year10_A42;
     }
 
@@ -46,11 +46,11 @@ public class DefaultAssumptions {
         Override_competitive_advantages_will_fade_A44 = override_competitive_advantages_will_fade_A44;
     }
 
-    public float getIf_yes_enter_capital_you_expect_after_year10_A45() {
+    public double getIf_yes_enter_capital_you_expect_after_year10_A45() {
         return If_yes_enter_capital_you_expect_after_year10_A45;
     }
 
-    public void setIf_yes_enter_capital_you_expect_after_year10_A45(float if_yes_enter_capital_you_expect_after_year10_A45) {
+    public void setIf_yes_enter_capital_you_expect_after_year10_A45(double if_yes_enter_capital_you_expect_after_year10_A45) {
         If_yes_enter_capital_you_expect_after_year10_A45 = if_yes_enter_capital_you_expect_after_year10_A45;
     }
 
@@ -62,11 +62,11 @@ public class DefaultAssumptions {
         Override_firm_has_no_chance_of_failure_A47 = override_firm_has_no_chance_of_failure_A47;
     }
 
-    public float getIf_yes_enter_the_probability_of_failure_A48() {
+    public double getIf_yes_enter_the_probability_of_failure_A48() {
         return If_yes_enter_the_probability_of_failure_A48;
     }
 
-    public void setIf_yes_enter_the_probability_of_failure_A48(float if_yes_enter_the_probability_of_failure_A48) {
+    public void setIf_yes_enter_the_probability_of_failure_A48(double if_yes_enter_the_probability_of_failure_A48) {
         If_yes_enter_the_probability_of_failure_A48 = if_yes_enter_the_probability_of_failure_A48;
     }
 
@@ -78,11 +78,11 @@ public class DefaultAssumptions {
         What_do_you_want_to_tie_your_proceeds_in_failure_to_A49 = what_do_you_want_to_tie_your_proceeds_in_failure_to_A49;
     }
 
-    public float getEnter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50() {
+    public double getEnter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50() {
         return Enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50;
     }
 
-    public void setEnter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50(float enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50) {
+    public void setEnter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50(double enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50) {
         Enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50 = enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50;
     }
 
@@ -102,11 +102,11 @@ public class DefaultAssumptions {
         Overrude_loss_tax_shield_assumption_A54 = overrude_loss_tax_shield_assumption_A54;
     }
 
-    public float getIf_yes_enter_NOL_carrying_into_year1_A55() {
+    public double getIf_yes_enter_NOL_carrying_into_year1_A55() {
         return If_yes_enter_NOL_carrying_into_year1_A55;
     }
 
-    public void setIf_yes_enter_NOL_carrying_into_year1_A55(float if_yes_enter_NOL_carrying_into_year1_A55) {
+    public void setIf_yes_enter_NOL_carrying_into_year1_A55(double if_yes_enter_NOL_carrying_into_year1_A55) {
         If_yes_enter_NOL_carrying_into_year1_A55 = if_yes_enter_NOL_carrying_into_year1_A55;
     }
 
@@ -118,11 +118,11 @@ public class DefaultAssumptions {
         Override_long_term_risk_free_assumptions_A57 = override_long_term_risk_free_assumptions_A57;
     }
 
-    public float getIf_yes_enter_riskfree_rate_after_year10_A58() {
+    public double getIf_yes_enter_riskfree_rate_after_year10_A58() {
         return If_yes_enter_riskfree_rate_after_year10_A58;
     }
 
-    public void setIf_yes_enter_riskfree_rate_after_year10_A58(float if_yes_enter_riskfree_rate_after_year10_A58) {
+    public void setIf_yes_enter_riskfree_rate_after_year10_A58(double if_yes_enter_riskfree_rate_after_year10_A58) {
         If_yes_enter_riskfree_rate_after_year10_A58 = if_yes_enter_riskfree_rate_after_year10_A58;
     }
 
@@ -134,11 +134,11 @@ public class DefaultAssumptions {
         Override_long_term_growth_rate_assumptions_A60 = override_long_term_growth_rate_assumptions_A60;
     }
 
-    public float getIf_yes_enter_the_growth_rate_in_perpetuity_A61() {
+    public double getIf_yes_enter_the_growth_rate_in_perpetuity_A61() {
         return If_yes_enter_the_growth_rate_in_perpetuity_A61;
     }
 
-    public void setIf_yes_enter_the_growth_rate_in_perpetuity_A61(float if_yes_enter_the_growth_rate_in_perpetuity_A61) {
+    public void setIf_yes_enter_the_growth_rate_in_perpetuity_A61(double if_yes_enter_the_growth_rate_in_perpetuity_A61) {
         If_yes_enter_the_growth_rate_in_perpetuity_A61 = if_yes_enter_the_growth_rate_in_perpetuity_A61;
     }
 
@@ -150,48 +150,48 @@ public class DefaultAssumptions {
         Override_cash_is_a_neutral_asset_A63 = override_cash_is_a_neutral_asset_A63;
     }
 
-    public float getIf_yes_enter_trapped_cash_or_entire_balance_A64() {
+    public double getIf_yes_enter_trapped_cash_or_entire_balance_A64() {
         return If_yes_enter_trapped_cash_or_entire_balance_A64;
     }
 
-    public void setIf_yes_enter_trapped_cash_or_entire_balance_A64(float if_yes_enter_trapped_cash_or_entire_balance_A64) {
+    public void setIf_yes_enter_trapped_cash_or_entire_balance_A64(double if_yes_enter_trapped_cash_or_entire_balance_A64) {
         If_yes_enter_trapped_cash_or_entire_balance_A64 = if_yes_enter_trapped_cash_or_entire_balance_A64;
     }
 
-    public float getAverage_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65() {
+    public double getAverage_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65() {
         return Average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65;
     }
 
-    public void setAverage_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65(float average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65) {
+    public void setAverage_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65(double average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65) {
         Average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65 = average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65;
     }
 
     private boolean Override_stable_growth_A41;
-    private float If_yes_enter_the_cost_of_capital_after_year10_A42;
+    private double If_yes_enter_the_cost_of_capital_after_year10_A42;
 
     private boolean Override_competitive_advantages_will_fade_A44;
-    private float If_yes_enter_capital_you_expect_after_year10_A45;
+    private double If_yes_enter_capital_you_expect_after_year10_A45;
 
     private boolean Override_firm_has_no_chance_of_failure_A47;
-    private float If_yes_enter_the_probability_of_failure_A48;
+    private double If_yes_enter_the_probability_of_failure_A48;
 
     private String What_do_you_want_to_tie_your_proceeds_in_failure_to_A49;   //  (v or B)
-    private float Enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50;
+    private double Enter_the_distress_proceeds_as_percentage_of_book_or_fair_value_A50;
 
     private boolean Override_long_term_tax_rate_assumption_A52;
 
     private boolean Overrude_loss_tax_shield_assumption_A54;
-    private float If_yes_enter_NOL_carrying_into_year1_A55;
+    private double If_yes_enter_NOL_carrying_into_year1_A55;
 
     private boolean Override_long_term_risk_free_assumptions_A57;
-    private float If_yes_enter_riskfree_rate_after_year10_A58;
+    private double If_yes_enter_riskfree_rate_after_year10_A58;
 
     private boolean Override_long_term_growth_rate_assumptions_A60;
-    private float If_yes_enter_the_growth_rate_in_perpetuity_A61;
+    private double If_yes_enter_the_growth_rate_in_perpetuity_A61;
 
     private boolean Override_cash_is_a_neutral_asset_A63;
-    private float If_yes_enter_trapped_cash_or_entire_balance_A64;
-    private float Average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65;
+    private double If_yes_enter_trapped_cash_or_entire_balance_A64;
+    private double Average_tax_rate_of_the_foreign_markets_where_the_cash_is_trapped_A65;
