+++ back_end/demo/src/main/java/com/back/example/InputSheet/MarketNumbers.java
@@ -3,29 +3,29 @@ package com.back.example.InputSheet;
 public class MarketNumbers {
 
 
-/*    public MarketNumbers(float riskfree_rate_A30, float initial_cost_of_capital_A31) {
+/*    public MarketNumbers(double riskfree_rate_A30, double initial_cost_of_capital_A31) {
         Riskfree_rate_A30 = riskfree_rate_A30;
         Initial_cost_of_capital_A31 = initial_cost_of_capital_A31;
     }*/
 
-    public float getRiskfree_rate_A30() {
+    public double getRiskfree_rate_A30() {
         return Riskfree_rate_A30;
     }
 
-    public void setRiskfree_rate_A30(float riskfree_rate_A30) {
+    public void setRiskfree_rate_A30(double riskfree_rate_A30) {
         Riskfree_rate_A30 = riskfree_rate_A30;
     }
 
-    public float getInitial_cost_of_capital_A31() {
+    public double getInitial_cost_of_capital_A31() {
         return Initial_cost_of_capital_A31;
     }
 
-    public void setInitial_cost_of_capital_A31(float initial_cost_of_capital_A31) {
+    public void setInitial_cost_of_capital_A31(double initial_cost_of_capital_A31) {
         Initial_cost_of_capital_A31 = initial_cost_of_capital_A31;
     }
 
-    private float Riskfree_rate_A30;
-    private float Initial_cost_of_capital_A31;
+    private double Riskfree_rate_A30;
+    private double Initial_cost_of_capital_A31;
