+++ back_end/demo/src/main/java/com/back/example/InputSheet/OtherInputs.java
@@ -3,7 +3,7 @@ package com.back.example.InputSheet;
 public class OtherInputs {
 
 
-/*    public OtherInputs(boolean do_you_have_employee_options_outstanding_A33, float number_of_options_outstanding_A34, float average_strike_price_A35, float average_maturity_A36, float standard_deviation_on_stock_price_A37) {
+/*    public OtherInputs(boolean do_you_have_employee_options_outstanding_A33, double number_of_options_outstanding_A34, double average_strike_price_A35, double average_maturity_A36, double standard_deviation_on_stock_price_A37) {
         Do_you_have_employee_options_outstanding_A33 = do_you_have_employee_options_outstanding_A33;
         Number_of_options_outstanding_A34 = number_of_options_outstanding_A34;
         Average_strike_price_A35 = average_strike_price_A35;
@@ -19,43 +19,43 @@ public class OtherInputs {
         Do_you_have_employee_options_outstanding_A33 = do_you_have_employee_options_outstanding_A33;
     }
 
-    public float getNumber_of_options_outstanding_A34() {
+    public double getNumber_of_options_outstanding_A34() {
         return Number_of_options_outstanding_A34;
     }
 
-    public void setNumber_of_options_outstanding_A34(float number_of_options_outstanding_A34) {
+    public void setNumber_of_options_outstanding_A34(double number_of_options_outstanding_A34) {
         Number_of_options_outstanding_A34 = number_of_options_outstanding_A34;
     }
 
-    public float getAverage_strike_price_A35() {
+    public double getAverage_strike_price_A35() {
         return Average_strike_price_A35;
     }
 
-    public void setAverage_strike_price_A35(float average_strike_price_A35) {
+    public void setAverage_strike_price_A35(double average_strike_price_A35) {
         Average_strike_price_A35 = average_strike_price_A35;
     }
 
-    public float getAverage_maturity_A36() {
+    public double getAverage_maturity_A36() {
         return Average_maturity_A36;
     }
 
-    public void setAverage_maturity_A36(float average_maturity_A36) {
+    public void setAverage_maturity_A36(double average_maturity_A36) {
         Average_maturity_A36 = average_maturity_A36;
     }
 
-    public float getStandard_deviation_on_stock_price_A37() {
+    public double getStandard_deviation_on_stock_price_A37() {
         return Standard_deviation_on_stock_price_A37;
     }
 
-    public void setStandard_deviation_on_stock_price_A37(float standard_deviation_on_stock_price_A37) {
+    public void setStandard_deviation_on_stock_price_A37(double standard_deviation_on_stock_price_A37) {
         Standard_deviation_on_stock_price_A37 = standard_deviation_on_stock_price_A37;
     }
 
     private boolean Do_you_have_employee_options_outstanding_A33;
-    private float Number_of_options_outstanding_A34;
-    private float Average_strike_price_A35;
-    private float Average_maturity_A36;
-    private float Standard_deviation_on_stock_price_A37;
+    private double Number_of_options_outstanding_A34;
+    private double Average_strike_price_A35;
+    private double Average_maturity_A36;
+    private double Standard_deviation_on_stock_price_A37;
 
 
 }
